+++ packages/mix-models/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/vueent/vueent/compare/v0.3.0...v0.4.0) (2023-05-26)
+
+
+### Features
+
+* validate mixin `pattern` options ([b15cf07](https://github.com/vueent/vueent/commit/b15cf0716c56c5bdd67c2c4b81c634203c65a4b1))
+
+
+
+
+
+# [0.3.0](https://github.com/vueent/vueent/compare/v0.2.0...v0.3.0) (2023-02-20)
+
+**Note:** Version bump only for package @vueent/mix-models
+
+
+
+
+
+# [0.2.0](https://github.com/vueent/vueent/compare/v0.1.3...v0.2.0) (2023-02-16)
+
+**Note:** Version bump only for package @vueent/mix-models
+
+
+
+
+
 ## [0.1.1](https://github.com/vueent/vueent/compare/v0.1.0...v0.1.1) (2023-01-19)
 
 **Note:** Version bump only for package @vueent/mix-models