+++ packages/reactive/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/vueent/vueent/compare/v0.3.0...v0.4.0) (2023-05-26)
+
+**Note:** Version bump only for package @vueent/reactive
+
+
+
+
+
+# [0.3.0](https://github.com/vueent/vueent/compare/v0.2.0...v0.3.0) (2023-02-20)
+
+**Note:** Version bump only for package @vueent/reactive
+
+
+
+
+
+# [0.2.0](https://github.com/vueent/vueent/compare/v0.1.3...v0.2.0) (2023-02-16)
+
+**Note:** Version bump only for package @vueent/reactive
+
+
+
+
+
 ## [0.0.16](https://github.com/vueent/vueent/compare/v0.0.15...v0.0.16) (2020-10-27)
 
 **Note:** Version bump only for package @vueent/reactive