+++ packages/store/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/vueent/vueent/compare/v0.3.0...v0.4.0) (2023-05-26)
+
+
+### Features
+
+* Protected `createInstance` wrapper for `Collection` class ([5b09505](https://github.com/vueent/vueent/commit/5b09505979d5d570befc050ca9bf41d7b79890bd))
+
+
+
+
+
+# [0.3.0](https://github.com/vueent/vueent/compare/v0.2.0...v0.3.0) (2023-02-20)
+
+**Note:** Version bump only for package @vueent/store
+
+
+
+
+
+# [0.2.0](https://github.com/vueent/vueent/compare/v0.1.3...v0.2.0) (2023-02-16)
+
+
+### Features
+
+* `force` option for collection's `find` and `findOne` methods ([6046197](https://github.com/vueent/vueent/commit/60461979845918f936343709fb34f4205b1e5c86))
+
+
+
+
+
+## [0.1.3](https://github.com/vueent/vueent/compare/v0.1.2...v0.1.3) (2023-02-13)
+
+**Note:** Version bump only for package @vueent/store
+
+
+
+
+
+## [0.1.2](https://github.com/vueent/vueent/compare/v0.1.1...v0.1.2) (2023-01-20)
+
+
+### Bug Fixes
+
+* Fix `@vueent/store` package building ([4c7c003](https://github.com/vueent/vueent/commit/4c7c00324f7f2e5592f685c63a5b7bd1e1e69db9)), closes [#47](https://github.com/vueent/vueent/issues/47)
+
+
+
+
+
 ## [0.1.1](https://github.com/vueent/vueent/compare/v0.1.0...v0.1.1) (2023-01-19)